+++ src/bin/words_to_data/main.rs
@@ -5,6 +5,7 @@ use clap::{Parser, Subcommand};
 
 mod build_dataset;
 mod convert_dataset;
+mod extract_changes;
 mod llm;
 mod match_amendments;
 mod score_amendments;
@@ -23,7 +24,9 @@ enum Command {
     BuildDataset(build_dataset::Args),
     /// Convert a serialized JSON dataset into a SQLite file
     ConvertDataset(convert_dataset::Args),
-    /// Extract amendment changes via an LLM and score them against the US Code diff
+    /// Extract word-level amendment changes for every bill via an LLM (writes into the dataset)
+    ExtractChanges(extract_changes::Args),
+    /// Score amendment changes against the US Code diff (deterministic, no LLM)
     ScoreAmendments(score_amendments::Args),
     /// Match bill amendments to US Code changes via an LLM and annotate the dataset
     MatchAmendments(match_amendments::Args),
@@ -33,6 +36,7 @@ fn main() {
     match Cli::parse().command {
         Command::BuildDataset(args) => build_dataset::run(args),
         Command::ConvertDataset(args) => convert_dataset::run(args),
+        Command::ExtractChanges(args) => extract_changes::run(args),
         Command::ScoreAmendments(args) => score_amendments::run(args),
         Command::MatchAmendments(args) => match_amendments::run(args),
     }
